% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/without-internet.R
\name{block_requests}
\alias{block_requests}
\title{Block HTTP requests}
\usage{
block_requests()
}
\value{
Nothing; called for its side effects.
}
\description{
This function intercepts HTTP requests made through \code{httr} and raises an
informative error instead. It is what \code{\link[=without_internet]{without_internet()}} does, minus the
automatic disabling of mocking when the context finishes.
}
\details{
Note that you in order to resume normal request behavior, you will need to
call \code{\link[=stop_mocking]{stop_mocking()}} yourself---this function does not clean up after itself
as 'without_internet` does.
}
\seealso{
\code{\link[=without_internet]{without_internet()}} \code{\link[=stop_mocking]{stop_mocking()}} \code{\link[=use_mock_api]{use_mock_api()}}
}
