% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture-requests.R
\name{capture_requests}
\alias{capture_requests}
\title{Collect API Responses as Mock Files}
\usage{
capture_requests(expr, path = ".")
}
\arguments{
\item{expr}{Code to run inside the context}

\item{path}{Where to save the mock files. Default is the current working
directory.}
}
\value{
The result of \code{expr}
}
\description{
This context collects the responses from requests you make and stores them as
mock files. This enables you to perform a series of requests against a live
server once and then build your test suite using those mocks, running your
tests in \code{with_mock_API}.
}
\examples{
\dontrun{
capture_requests({
    GET("http://httpbin.org/get")
    GET("http://httpbin.org")
    GET("http://httpbin.org/response-headers",
        query=list(`Content-Type`="application/json"))
    utils::download.file("http://httpbin.org/gzip", tempfile())
})
}
}
