% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette.R
\name{change_state}
\alias{change_state}
\title{Handle a change of server state}
\usage{
change_state()
}
\value{
Invisibly, the return of \code{.mockPaths()} with the new path added.
}
\description{
Changes to server state can be handled by setting a new \code{\link[=.mockPaths]{.mockPaths()}}. In
vignettes, these are handled by creating subdirectories with integer names.
This function increments that integer subdirectory and adds it to
\code{.mockPaths()} so that changed server state can be captured and loaded.
}
\details{
In a vignette,
put a call to \code{change_state()} before any code block that makes a change on
the server, or rather, before any code block that might repeat the same
request previously done and expect a different result.
}
