% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock-api.R
\name{find_mock_file}
\alias{find_mock_file}
\title{Go through mock paths to find the local mock file location}
\usage{
find_mock_file(file)
}
\arguments{
\item{file}{A file path, as generated by \code{\link[=build_mock_url]{build_mock_url()}}.}
}
\value{
A path to a file that exists, or \code{NULL} if none found.
}
\description{
Go through mock paths to find the local mock file location
}
\keyword{internal}
