% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-redactor.R
\name{get_current_redactor}
\alias{get_current_redactor}
\title{Fetch the active redacting function}
\usage{
get_current_redactor()
}
\value{
A redacting function.
}
\description{
Called inside \code{\link[=capture_requests]{capture_requests()}}. If using the default redactor, it checks
each time it is called to see if any new packages have been attached, in case
there are package redactors in them.
}
\keyword{internal}
