% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-requester.R
\name{get_current_requester}
\alias{get_current_requester}
\title{Fetch the active requesting function}
\usage{
get_current_requester()
}
\value{
A requesting function.
}
\description{
Called inside \code{\link[=capture_requests]{capture_requests()}}. If using the default requester, it checks
each time it is called to see if any new packages have been attached, in case
there are package requesters in them.
}
\seealso{
\code{\link[=get_current_redactor]{get_current_redactor()}}
}
\keyword{internal}
