% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redact.R
\name{gsub_response}
\alias{gsub_response}
\title{Find and replace within a 'response'}
\usage{
gsub_response(response, pattern, replacement, ...)
}
\arguments{
\item{response}{An 'httr' \code{response} object to sanitize.}

\item{pattern}{From \code{\link[base:gsub]{base::gsub()}}: "character string containing a regular
expression (or character string for \code{fixed = TRUE}) to be matched in the
given character vector." Passed to \code{gsub()}. See the docs for \code{gsub()} for
further details.}

\item{replacement}{A replacement for the matched pattern, possibly including
regular expression backreferences. Passed to \code{gsub()}. See the docs for
\code{gsub()} for further details.}

\item{...}{Additional logical arguments passed to \code{gsub()}: \code{ignore.case},
\code{perl}, \code{fixed}, and \code{useBytes} are the possible options.}
}
\value{
A \code{response} object with the pattern replaced in the URLs and
response body.
}
\description{
This function passes its arguments to \code{\link[base:gsub]{base::gsub()}} in order to find and
replace string patterns (regular expressions) in three attributes of an
\code{httr} 'response' object: (1) the response body; (2) the response URL; and
(3) the request URL (that is, the URL appears twice in the response object).
}
\details{
Note that, unlike \code{gsub()}, the first argument of the function is \code{response},
not \code{pattern}, while the equivalent argument in \code{gsub()}, "\code{x}", is placed
third. This difference is to maintain consistency with the other redactor
functions in \code{httptest}, which all take \code{response} as the first argument.
}
