% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{public}
\alias{public}
\title{Test that functions are exported}
\usage{
public(...)
}
\arguments{
\item{...}{Code to evaluate}
}
\value{
The result of \code{...} evaluated in the global environment (and not
the package environment).
}
\description{
It's easy to forget to document and export a new function. Wrap 'public()'
around test blocks to assert that the functions they call
are exported (and thus fail if you haven't documented them with @export)
}
