% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-requester.R
\name{set_requester}
\alias{set_requester}
\title{Set a request preprocessor}
\usage{
set_requester(FUN)
}
\arguments{
\item{FUN}{A function that modifies \code{request} objects}
}
\value{
Invisibly, the return of the call to \code{options()}
}
\description{
Set a request preprocessor
}
