% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette.R
\name{start_vignette}
\alias{start_vignette}
\alias{end_vignette}
\title{Set mocking/capturing state for a vignette}
\usage{
start_vignette(...)

end_vignette()
}
\arguments{
\item{...}{Optional arguments passed to \code{start_capturing()}}
}
\value{
Nothing; called for its side effect of starting/ending
response recording or mocking.
}
\description{
Use \code{start_vignette()} to either use previously recorded responses or capture
real responses for future use, depending on the value of the \code{RECORD}
environment variable.
}
\details{
In a vignette or other R Markdown or Sweave document, place
\code{start_vignette()} in an R code block at the beginning,
before the first API request is made, and put
\code{end_vignette()} in a R code chunk at the end. You may
want to make those R code chunks have \code{echo=FALSE} in order to hide the fact
that you're calling them.

The environment variable \code{RECORD} determines whether you're making real
requests and capturing them or whether you're loading previously recorded
mocks. If \code{RECORD} is \code{true} (or \code{TRUE}: it's case insensitive),
\code{start_vignette()} calls \code{\link[=start_capturing]{start_capturing()}}. Otherwise (by default), it
calls \code{\link[=use_mock_API]{use_mock_API()}}.
}
\seealso{
\code{\link[=start_capturing]{start_capturing()}} \code{\link[=use_mock_API]{use_mock_API()}} \code{vignette("vignettes", package="httptest")}
}
