% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace.R
\name{stop_mocking}
\alias{stop_mocking}
\title{Turn off request mocking}
\usage{
stop_mocking()
}
\value{
Nothing; called for its side effects
}
\description{
This function "untraces" the \code{httr} request functions so that normal, real
requesting behavior can be resumed.
}
