% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-httptest.R
\name{use_httptest}
\alias{use_httptest}
\title{Use 'httptest' in your tests}
\usage{
use_httptest(path = ".")
}
\arguments{
\item{path}{character path to the package}
}
\value{
Nothing: called for file system side effects.
}
\description{
This function adds \code{httptest} to Suggests in the package DESCRIPTION and
loads it in \code{tests/testthat/helper.R}. Call it once when you're setting up
a new package test suite.
}
\details{
The function is idempotent: if \code{httptest} is already added to these files, no
additional changes will be made.
}
