% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock-api.R
\name{use_mock_api}
\alias{use_mock_api}
\title{Turn on API mocking}
\usage{
use_mock_api()
}
\value{
Nothing; called for its side effects.
}
\description{
This function intercepts HTTP requests made through \code{httr} and serves mock
file responses instead. It is what \code{\link[=with_mock_api]{with_mock_api()}} does, minus the
automatic disabling of mocking when the context finishes.
}
\details{
Note that you in order to resume normal request behavior, you will need to
call \code{\link[=stop_mocking]{stop_mocking()}} yourself---this function does not clean up after itself
as \code{with_mock_api} does.
}
\seealso{
\code{\link[=with_mock_api]{with_mock_api()}} \code{\link[=stop_mocking]{stop_mocking()}} \code{\link[=block_requests]{block_requests()}}
}
