% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect-header.R
\name{with_trace}
\alias{with_trace}
\title{Wrapper around 'trace' to untrace when finished}
\usage{
with_trace(x, where = topenv(parent.frame()), print = FALSE, ..., expr)
}
\arguments{
\item{x}{Name of function to trace. See \code{\link[base:trace]{base::trace()}}.}

\item{where}{where to look for the function to be traced.}

\item{print}{Logical: print a message every time the traced function is hit?
Default is \code{FALSE}; note that in \code{trace}, the default is \code{TRUE}.}

\item{...}{Additional arguments to pass to \code{trace}. At minimum, must include
either \code{tracer} and \code{at}, or \code{exit}.}

\item{expr}{Code to run inside the context}
}
\value{
The result of \code{expr}
}
\description{
Wrapper around 'trace' to untrace when finished
}
