% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contexts.R
\name{without_internet}
\alias{without_internet}
\title{Make all HTTP requests raise an error}
\usage{
without_internet(expr)
}
\arguments{
\item{expr}{Code to run inside the mock context}
}
\value{
The result of \code{expr}
}
\description{
Make all HTTP requests raise an error
}

