% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect-request.R
\name{expect-verb}
\alias{expect-verb}
\alias{expect_DELETE}
\alias{expect_GET}
\alias{expect_PATCH}
\alias{expect_POST}
\alias{expect_PUT}
\alias{expect_no_request}
\title{Expecations for mocked HTTP requests}
\usage{
expect_GET(object, url = "", ...)

expect_POST(object, url = "", ...)

expect_PATCH(object, url = "", ...)

expect_PUT(object, url = "", ...)

expect_DELETE(object, url = "")

expect_no_request(object, ...)
}
\arguments{
\item{object}{Code to execute that may cause an HTTP request}

\item{url}{character: the URL you expect a request to be made to}

\item{...}{character segments of a request payload you expect to be included
in the request body, to be joined together by \code{paste0}}
}
\value{
A \code{testthat} 'expectation'.
}
\description{
Use these inside one of the mock contexts.
}
\examples{
without_internet({
    expect_GET(httr::GET("http://httpbin.org/get"),
        "http://httpbin.org/get")
    expect_no_request(rnorm(5))
})
}

